+++ src/internal/m365/service/groups/export_test.go
@@ -20,6 +20,7 @@ import (
 	"github.com/alcionai/corso/src/pkg/control"
 	"github.com/alcionai/corso/src/pkg/export"
 	"github.com/alcionai/corso/src/pkg/fault"
+	"github.com/alcionai/corso/src/pkg/metrics"
 	"github.com/alcionai/corso/src/pkg/path"
 	"github.com/alcionai/corso/src/pkg/services/m365/api"
 )
@@ -95,7 +96,7 @@ func (suite *ExportUnitSuite) TestExportRestoreCollections_messages() {
 		},
 	}
 
-	stats := data.ExportStats{}
+	stats := metrics.ExportStats{}
 
 	ecs, err := NewGroupsHandler(control.DefaultOptions(), api.Client{}, nil).
 		ProduceExportCollections(
@@ -129,7 +130,7 @@ func (suite *ExportUnitSuite) TestExportRestoreCollections_messages() {
 
 	assert.Equal(t, expectedItems, fitems, "items")
 
-	expectedStats := data.ExportStats{}
+	expectedStats := metrics.ExportStats{}
 	expectedStats.UpdateBytes(path.ChannelMessagesCategory, int64(size))
 	expectedStats.UpdateResourceCount(path.ChannelMessagesCategory)
 	assert.Equal(t, expectedStats, stats, "stats")
@@ -200,7 +201,7 @@ func (suite *ExportUnitSuite) TestExportRestoreCollections_libraries() {
 	handler := NewGroupsHandler(control.DefaultOptions(), api.Client{}, nil)
 	handler.CacheItemInfo(dii)
 
-	stats := data.ExportStats{}
+	stats := metrics.ExportStats{}
 
 	ecs, err := handler.ProduceExportCollections(
 		ctx,
@@ -232,7 +233,7 @@ func (suite *ExportUnitSuite) TestExportRestoreCollections_libraries() {
 
 	assert.Equal(t, expectedItems, fitems, "items")
 
-	expectedStats := data.ExportStats{}
+	expectedStats := metrics.ExportStats{}
 	expectedStats.UpdateBytes(path.FilesCategory, int64(size))
 	expectedStats.UpdateResourceCount(path.FilesCategory)
 	assert.Equal(t, expectedStats, stats, "stats")
