+++ src/internal/m365/service/sharepoint/export_test.go
@@ -19,6 +19,7 @@ import (
 	"github.com/alcionai/corso/src/pkg/control"
 	"github.com/alcionai/corso/src/pkg/export"
 	"github.com/alcionai/corso/src/pkg/fault"
+	"github.com/alcionai/corso/src/pkg/metrics"
 	"github.com/alcionai/corso/src/pkg/path"
 	"github.com/alcionai/corso/src/pkg/services/m365/api"
 )
@@ -129,7 +130,7 @@ func (suite *ExportUnitSuite) TestExportRestoreCollections() {
 			handler := NewSharePointHandler(control.DefaultOptions(), api.Client{}, nil)
 			handler.CacheItemInfo(test.itemInfo)
 
-			stats := data.ExportStats{}
+			stats := metrics.ExportStats{}
 
 			ecs, err := handler.ProduceExportCollections(
 				ctx,
@@ -160,7 +161,7 @@ func (suite *ExportUnitSuite) TestExportRestoreCollections() {
 
 			assert.Equal(t, expectedItems, fitems, "items")
 
-			expectedStats := data.ExportStats{}
+			expectedStats := metrics.ExportStats{}
 			expectedStats.UpdateBytes(path.FilesCategory, int64(size))
 			expectedStats.UpdateResourceCount(path.FilesCategory)
 			assert.Equal(t, expectedStats, stats, "stats")
